--- prototype/frontend/src/components/__tests__/BattleConfigTab.pbt.test.tsx
@@ -1,881 +1,0 @@
-import { describe, it, expect, vi, afterEach } from 'vitest';
-import { render, screen, cleanup } from '@testing-library/react';
-import * as fc from 'fast-check';
-import BattleConfigTab from '../BattleConfigTab';
-
-// Cleanup after each test
-afterEach(() => {
-  cleanup();
-  vi.clearAllMocks();
-});
-
-/**
- * Property-Based Tests for Robot Detail Page Visual Enhancement
- * Feature: robot-detail-page-visual-enhancement
- * Testing Framework: fast-check with minimum 100 iterations
- */
-
-/**
- * Property 18: HP and Shield Bar Color Coding
- * **Validates: Requirements 6.6**
- * 
- * For any robot, when HP is 70-100%, the HP bar should be green (#3fb950);
- * when 30-69%, amber (#d29922); when 1-29%, red (#f85149).
- * 
- * This property test verifies that across all valid HP percentages (0-100%),
- * the HP bar color is correctly applied based on the specified thresholds.
- * The same color coding applies to shield bars.
- */
-describe('Property 18: HP and Shield Bar Color Coding (Property-Based Test)', () => {
-  // Mock functions for props
-  const mockOnRobotUpdate = vi.fn();
-  const mockOnEquipWeapon = vi.fn();
-  const mockOnUnequipWeapon = vi.fn();
-
-  // Helper function to create a robot with specific HP/shield percentages
-  const createRobot = (currentHP: number, maxHP: number, currentShield: number, maxShield: number) => ({
-    id: 1,
-    name: 'Test Robot',
-    currentHP,
-    maxHP,
-    currentShield,
-    maxShield,
-    battleReadiness: 100,
-    repairCost: 0,
-    loadoutType: 'single',
-    mainWeaponId: null,
-    offhandWeaponId: null,
-    stance: 'balanced',
-    yieldThreshold: 50,
-    mainWeapon: null,
-    offhandWeapon: null,
-  });
-
-  it('should display green HP bar (#3fb950) when HP is 70-100%', () => {
-    fc.assert(
-      fc.property(
-        // Generate HP percentage in the 70-100% range (avoid exact 70 due to floating point)
-        fc.float({ min: Math.fround(70.01), max: 100, noNaN: true }),
-        // Generate max HP value
-        fc.integer({ min: 100, max: 10000 }),
-        (hpPercent, maxHP) => {
-          // Calculate current HP based on percentage
-          const currentHP = Math.floor((hpPercent / 100) * maxHP);
-          
-          // Create robot with calculated HP values
-          const robot = createRobot(currentHP, maxHP, maxHP, maxHP);
-          
-          const { container: _container } = render(
-            <BattleConfigTab
-              robot={robot}
-              onRobotUpdate={mockOnRobotUpdate}
-              onEquipWeapon={mockOnEquipWeapon}
-              onUnequipWeapon={mockOnUnequipWeapon}
-            />
-          );
-
-          // Property: HP bar should have green color (#3fb950)
-          // Find the HP bar by looking for the progress bar div inside the HP section
-          const hullIntegrityTexts = screen.getAllByText('Hull Integrity');
-          const hpSection = hullIntegrityTexts[0].parentElement!.parentElement;
-          expect(hpSection).toBeTruthy();
-          
-          // Find the colored bar within the HP section
-          const hpBar = hpSection!.querySelector('.h-full') as HTMLElement;
-          expect(hpBar).toBeTruthy();
-          
-          const backgroundColor = hpBar.style.backgroundColor;
-          
-          // Convert hex to rgb for comparison (browsers may return rgb format)
-          // #3fb950 = rgb(63, 185, 80)
-          const isGreen = backgroundColor === 'rgb(63, 185, 80)' || 
-                         backgroundColor === '#3fb950';
-          
-          expect(isGreen).toBe(true);
-          
-          // Cleanup
-          cleanup();
-        }
-      ),
-      { numRuns: 100 } // Minimum 100 iterations as specified in design
-    );
-  });
-
-  it('should display amber HP bar (#d29922) when HP is 30-69%', () => {
-    fc.assert(
-      fc.property(
-        // Generate HP percentage in the 30-69% range (avoid exact boundaries)
-        fc.float({ min: Math.fround(30.01), max: Math.fround(69.98), noNaN: true }),
-        // Generate max HP value
-        fc.integer({ min: 100, max: 10000 }),
-        (hpPercent, maxHP) => {
-          // Calculate current HP based on percentage
-          const currentHP = Math.floor((hpPercent / 100) * maxHP);
-          
-          // Create robot with calculated HP values
-          const robot = createRobot(currentHP, maxHP, maxHP, maxHP);
-          
-          const { container: _container } = render(
-            <BattleConfigTab
-              robot={robot}
-              onRobotUpdate={mockOnRobotUpdate}
-              onEquipWeapon={mockOnEquipWeapon}
-              onUnequipWeapon={mockOnUnequipWeapon}
-            />
-          );
-
-          // Property: HP bar should have amber color (#d29922)
-          const hullIntegrityTexts = screen.getAllByText('Hull Integrity');
-          const hpSection = hullIntegrityTexts[0].parentElement!.parentElement;
-          expect(hpSection).toBeTruthy();
-          
-          // Find the colored bar within the HP section
-          const hpBar = hpSection!.querySelector('.h-full') as HTMLElement;
-          expect(hpBar).toBeTruthy();
-          
-          const backgroundColor = hpBar.style.backgroundColor;
-          
-          // Convert hex to rgb for comparison
-          // #d29922 = rgb(210, 153, 34)
-          const isAmber = backgroundColor === 'rgb(210, 153, 34)' || 
-                         backgroundColor === '#d29922';
-          
-          expect(isAmber).toBe(true);
-          
-          // Cleanup
-          cleanup();
-        }
-      ),
-      { numRuns: 100 }
-    );
-  });
-
-  it('should display red HP bar (#f85149) when HP is 1-29%', () => {
-    fc.assert(
-      fc.property(
-        // Generate HP percentage in the 1-29% range
-        fc.float({ min: 1, max: Math.fround(29.99), noNaN: true }),
-        // Generate max HP value
-        fc.integer({ min: 100, max: 10000 }),
-        (hpPercent, maxHP) => {
-          // Calculate current HP based on percentage
-          const currentHP = Math.max(1, Math.floor((hpPercent / 100) * maxHP));
-          
-          // Create robot with calculated HP values
-          const robot = createRobot(currentHP, maxHP, maxHP, maxHP);
-          
-          const { container: _container } = render(
-            <BattleConfigTab
-              robot={robot}
-              onRobotUpdate={mockOnRobotUpdate}
-              onEquipWeapon={mockOnEquipWeapon}
-              onUnequipWeapon={mockOnUnequipWeapon}
-            />
-          );
-
-          // Property: HP bar should have red color (#f85149)
-          const hullIntegrityTexts = screen.getAllByText('Hull Integrity');
-          const hpSection = hullIntegrityTexts[0].parentElement!.parentElement;
-          expect(hpSection).toBeTruthy();
-          
-          // Find the colored bar within the HP section
-          const hpBar = hpSection!.querySelector('.h-full') as HTMLElement;
-          expect(hpBar).toBeTruthy();
-          
-          const backgroundColor = hpBar.style.backgroundColor;
-          
-          // Convert hex to rgb for comparison
-          // #f85149 = rgb(248, 81, 73)
-          const isRed = backgroundColor === 'rgb(248, 81, 73)' || 
-                       backgroundColor === '#f85149';
-          
-          expect(isRed).toBe(true);
-          
-          // Cleanup
-          cleanup();
-        }
-      ),
-      { numRuns: 100 }
-    );
-  });
-
-  it('should display correct color for edge case HP percentages', () => {
-    fc.assert(
-      fc.property(
-        // Generate edge case HP percentages
-        fc.constantFrom(
-          { percent: 100, expectedColor: '#3fb950', colorName: 'green' },
-          { percent: 70, expectedColor: '#3fb950', colorName: 'green' },
-          { percent: 69.99, expectedColor: '#d29922', colorName: 'amber' },
-          { percent: 30, expectedColor: '#d29922', colorName: 'amber' },
-          { percent: 29.99, expectedColor: '#f85149', colorName: 'red' },
-          { percent: 1, expectedColor: '#f85149', colorName: 'red' }
-        ),
-        // Generate max HP value
-        fc.integer({ min: 100, max: 10000 }),
-        (scenario, maxHP) => {
-          const { percent, expectedColor } = scenario;
-          
-          // Calculate current HP based on percentage
-          const currentHP = Math.max(1, Math.floor((percent / 100) * maxHP));
-          
-          // Create robot with calculated HP values
-          const robot = createRobot(currentHP, maxHP, maxHP, maxHP);
-          
-          const { container: _container } = render(
-            <BattleConfigTab
-              robot={robot}
-              onRobotUpdate={mockOnRobotUpdate}
-              onEquipWeapon={mockOnEquipWeapon}
-              onUnequipWeapon={mockOnUnequipWeapon}
-            />
-          );
-
-          // Property: HP bar should have the expected color for edge cases
-          const hullIntegrityTexts = screen.getAllByText('Hull Integrity');
-          const hpSection = hullIntegrityTexts[0].parentElement!.parentElement;
-          expect(hpSection).toBeTruthy();
-          
-          // Find the colored bar within the HP section
-          const hpBar = hpSection!.querySelector('.h-full') as HTMLElement;
-          expect(hpBar).toBeTruthy();
-          
-          const backgroundColor = hpBar.style.backgroundColor;
-          
-          // Convert expected hex to rgb
-          const hexToRgb = (hex: string) => {
-            const r = parseInt(hex.slice(1, 3), 16);
-            const g = parseInt(hex.slice(3, 5), 16);
-            const b = parseInt(hex.slice(5, 7), 16);
-            return `rgb(${r}, ${g}, ${b})`;
-          };
-          
-          const expectedRgb = hexToRgb(expectedColor);
-          const hasCorrectColor = backgroundColor === expectedRgb || 
-                                 backgroundColor === expectedColor;
-          
-          expect(hasCorrectColor).toBe(true);
-          
-          // Cleanup
-          cleanup();
-        }
-      ),
-      { numRuns: 100 }
-    );
-  });
-
-  it('should display green shield bar (#3fb950) when shield is 70-100%', () => {
-    fc.assert(
-      fc.property(
-        // Generate shield percentage in the 70-100% range (avoid exact 70)
-        fc.float({ min: Math.fround(70.01), max: 100, noNaN: true }),
-        // Generate max shield value
-        fc.integer({ min: 100, max: 10000 }),
-        (shieldPercent, maxShield) => {
-          // Calculate current shield based on percentage
-          const currentShield = Math.floor((shieldPercent / 100) * maxShield);
-          
-          // Create robot with full HP and calculated shield values
-          const robot = createRobot(1000, 1000, currentShield, maxShield);
-          
-          const { container: _container } = render(
-            <BattleConfigTab
-              robot={robot}
-              onRobotUpdate={mockOnRobotUpdate}
-              onEquipWeapon={mockOnEquipWeapon}
-              onUnequipWeapon={mockOnUnequipWeapon}
-            />
-          );
-
-          // Property: Shield bar should have green color (#3fb950)
-          const energyShieldTexts = screen.getAllByText('Energy Shield');
-          const shieldSection = energyShieldTexts[0].parentElement!.parentElement;
-          expect(shieldSection).toBeTruthy();
-          
-          // Find the colored bar within the shield section
-          const shieldBar = shieldSection!.querySelector('.h-full') as HTMLElement;
-          expect(shieldBar).toBeTruthy();
-          
-          const backgroundColor = shieldBar.style.backgroundColor;
-          
-          // #3fb950 = rgb(63, 185, 80)
-          const isGreen = backgroundColor === 'rgb(63, 185, 80)' || 
-                         backgroundColor === '#3fb950';
-          
-          expect(isGreen).toBe(true);
-          
-          // Cleanup
-          cleanup();
-        }
-      ),
-      { numRuns: 100 }
-    );
-  });
-
-  it('should display amber shield bar (#d29922) when shield is 30-69%', () => {
-    fc.assert(
-      fc.property(
-        // Generate shield percentage in the 30-69% range (avoid exact boundaries)
-        fc.float({ min: Math.fround(30.01), max: Math.fround(69.98), noNaN: true }),
-        // Generate max shield value
-        fc.integer({ min: 100, max: 10000 }),
-        (shieldPercent, maxShield) => {
-          // Calculate current shield based on percentage
-          const currentShield = Math.floor((shieldPercent / 100) * maxShield);
-          
-          // Create robot with full HP and calculated shield values
-          const robot = createRobot(1000, 1000, currentShield, maxShield);
-          
-          const { container: _container } = render(
-            <BattleConfigTab
-              robot={robot}
-              onRobotUpdate={mockOnRobotUpdate}
-              onEquipWeapon={mockOnEquipWeapon}
-              onUnequipWeapon={mockOnUnequipWeapon}
-            />
-          );
-
-          // Property: Shield bar should have amber color (#d29922)
-          const energyShieldTexts = screen.getAllByText('Energy Shield');
-          const shieldSection = energyShieldTexts[0].parentElement!.parentElement;
-          expect(shieldSection).toBeTruthy();
-          
-          // Find the colored bar within the shield section
-          const shieldBar = shieldSection!.querySelector('.h-full') as HTMLElement;
-          expect(shieldBar).toBeTruthy();
-          
-          const backgroundColor = shieldBar.style.backgroundColor;
-          
-          // #d29922 = rgb(210, 153, 34)
-          const isAmber = backgroundColor === 'rgb(210, 153, 34)' || 
-                         backgroundColor === '#d29922';
-          
-          expect(isAmber).toBe(true);
-          
-          // Cleanup
-          cleanup();
-        }
-      ),
-      { numRuns: 100 }
-    );
-  });
-
-  it('should display red shield bar (#f85149) when shield is 1-29%', () => {
-    fc.assert(
-      fc.property(
-        // Generate shield percentage in the 1-29% range (avoid exact 30)
-        fc.float({ min: 1, max: Math.fround(29.98), noNaN: true }),
-        // Generate max shield value
-        fc.integer({ min: 100, max: 10000 }),
-        (shieldPercent, maxShield) => {
-          // Calculate current shield based on percentage
-          const currentShield = Math.max(1, Math.floor((shieldPercent / 100) * maxShield));
-          
-          // Create robot with full HP and calculated shield values
-          const robot = createRobot(1000, 1000, currentShield, maxShield);
-          
-          const { container: _container } = render(
-            <BattleConfigTab
-              robot={robot}
-              onRobotUpdate={mockOnRobotUpdate}
-              onEquipWeapon={mockOnEquipWeapon}
-              onUnequipWeapon={mockOnUnequipWeapon}
-            />
-          );
-
-          // Property: Shield bar should have red color (#f85149)
-          const energyShieldTexts = screen.getAllByText('Energy Shield');
-          const shieldSection = energyShieldTexts[0].parentElement!.parentElement;
-          expect(shieldSection).toBeTruthy();
-          
-          // Find the colored bar within the shield section
-          const shieldBar = shieldSection!.querySelector('.h-full') as HTMLElement;
-          expect(shieldBar).toBeTruthy();
-          
-          const backgroundColor = shieldBar.style.backgroundColor;
-          
-          // #f85149 = rgb(248, 81, 73)
-          const isRed = backgroundColor === 'rgb(248, 81, 73)' || 
-                       backgroundColor === '#f85149';
-          
-          expect(isRed).toBe(true);
-          
-          // Cleanup
-          cleanup();
-        }
-      ),
-      { numRuns: 100 }
-    );
-  });
-
-  it('should apply consistent color coding to both HP and shield bars', () => {
-    fc.assert(
-      fc.property(
-        // Generate same percentage for both HP and shield
-        fc.float({ min: 1, max: 100, noNaN: true }),
-        // Generate max values
-        fc.integer({ min: 100, max: 10000 }),
-        fc.integer({ min: 100, max: 10000 }),
-        (percent, maxHP, maxShield) => {
-          // Calculate current values based on percentage
-          const currentHP = Math.max(1, Math.floor((percent / 100) * maxHP));
-          const currentShield = Math.max(1, Math.floor((percent / 100) * maxShield));
-          
-          // Create robot with calculated values
-          const robot = createRobot(currentHP, maxHP, currentShield, maxShield);
-          
-          const { container: _container } = render(
-            <BattleConfigTab
-              robot={robot}
-              onRobotUpdate={mockOnRobotUpdate}
-              onEquipWeapon={mockOnEquipWeapon}
-              onUnequipWeapon={mockOnUnequipWeapon}
-            />
-          );
-
-          // Property: Both bars should have the same color when at the same percentage
-          const hullIntegrityTexts = screen.getAllByText('Hull Integrity');
-          const hpSection = hullIntegrityTexts[0].parentElement!.parentElement;
-          expect(hpSection).toBeTruthy();
-          
-          const energyShieldTexts = screen.getAllByText('Energy Shield');
-          const shieldSection = energyShieldTexts[0].parentElement!.parentElement;
-          expect(shieldSection).toBeTruthy();
-          
-          const hpBar = hpSection!.querySelector('.h-full') as HTMLElement;
-          const shieldBar = shieldSection!.querySelector('.h-full') as HTMLElement;
-          
-          expect(hpBar).toBeTruthy();
-          expect(shieldBar).toBeTruthy();
-          
-          const hpColor = hpBar.style.backgroundColor;
-          const shieldColor = shieldBar.style.backgroundColor;
-          
-          // Both should have the same color
-          expect(hpColor).toBe(shieldColor);
-          
-          // Verify the color matches the expected threshold
-          if (percent >= 70) {
-            const isGreen = hpColor === 'rgb(63, 185, 80)' || hpColor === '#3fb950';
-            expect(isGreen).toBe(true);
-          } else if (percent >= 30) {
-            const isAmber = hpColor === 'rgb(210, 153, 34)' || hpColor === '#d29922';
-            expect(isAmber).toBe(true);
-          } else {
-            const isRed = hpColor === 'rgb(248, 81, 73)' || hpColor === '#f85149';
-            expect(isRed).toBe(true);
-          }
-          
-          // Cleanup
-          cleanup();
-        }
-      ),
-      { numRuns: 100 }
-    );
-  });
-
-  it('should maintain color coding across different max HP and shield values', () => {
-    fc.assert(
-      fc.property(
-        // Generate random HP percentage
-        fc.float({ min: 1, max: 100, noNaN: true }),
-        // Generate various max HP values
-        fc.integer({ min: 100, max: 50000 }),
-        (hpPercent, maxHP) => {
-          // Calculate current HP based on percentage
-          const currentHP = Math.max(1, Math.floor((hpPercent / 100) * maxHP));
-          
-          // Create robot with calculated HP values
-          const robot = createRobot(currentHP, maxHP, maxHP, maxHP);
-          
-          const { container: _container } = render(
-            <BattleConfigTab
-              robot={robot}
-              onRobotUpdate={mockOnRobotUpdate}
-              onEquipWeapon={mockOnEquipWeapon}
-              onUnequipWeapon={mockOnUnequipWeapon}
-            />
-          );
-
-          // Property: Color should be determined by percentage, not absolute values
-          const hullIntegrityTexts = screen.getAllByText('Hull Integrity');
-          const hpSection = hullIntegrityTexts[0].parentElement!.parentElement;
-          expect(hpSection).toBeTruthy();
-          
-          const hpBar = hpSection!.querySelector('.h-full') as HTMLElement;
-          expect(hpBar).toBeTruthy();
-          
-          const backgroundColor = hpBar.style.backgroundColor;
-          
-          // Verify color matches percentage threshold regardless of max HP
-          if (hpPercent >= 70) {
-            const isGreen = backgroundColor === 'rgb(63, 185, 80)' || 
-                           backgroundColor === '#3fb950';
-            expect(isGreen).toBe(true);
-          } else if (hpPercent >= 30) {
-            const isAmber = backgroundColor === 'rgb(210, 153, 34)' || 
-                           backgroundColor === '#d29922';
-            expect(isAmber).toBe(true);
-          } else {
-            const isRed = backgroundColor === 'rgb(248, 81, 73)' || 
-                         backgroundColor === '#f85149';
-            expect(isRed).toBe(true);
-          }
-          
-          // Cleanup
-          cleanup();
-        }
-      ),
-      { numRuns: 100 }
-    );
-  });
-});
-
-/**
- * Property 19: Configuration Change Feedback
- * **Validates: Requirements 6.8**
- * 
- * For any configuration change (loadout, stance, yield), the effective stats display
- * should update immediately to reflect the new values.
- * 
- * This property test verifies that when configuration changes occur, the component
- * properly triggers updates that would cause the effective stats to recalculate.
- * We test this by verifying that the onRobotUpdate callback is called with the
- * correct new configuration values, which would trigger a re-render with updated stats.
- */
-describe('Property 19: Configuration Change Feedback (Property-Based Test)', () => {
-  const mockOnRobotUpdate = vi.fn();
-  const mockOnEquipWeapon = vi.fn();
-  const mockOnUnequipWeapon = vi.fn();
-
-  // Helper function to create a robot with specific configuration
-  const createRobot = (loadoutType: string, stance: string, yieldThreshold: number) => ({
-    id: 1,
-    name: 'Test Robot',
-    currentHP: 1000,
-    maxHP: 1000,
-    currentShield: 500,
-    maxShield: 500,
-    battleReadiness: 100,
-    repairCost: 0,
-    loadoutType,
-    mainWeaponId: null,
-    offhandWeaponId: null,
-    stance,
-    yieldThreshold,
-    mainWeapon: null,
-    offhandWeapon: null,
-  });
-
-  it('should trigger update callback when loadout type changes', () => {
-    fc.assert(
-      fc.property(
-        // Generate initial loadout type
-        fc.constantFrom('single', 'weapon_shield', 'two_handed', 'dual_wield'),
-        // Generate new loadout type (different from initial)
-        fc.constantFrom('single', 'weapon_shield', 'two_handed', 'dual_wield'),
-        // Generate stance
-        fc.constantFrom('offensive', 'defensive', 'balanced'),
-        // Generate yield threshold
-        fc.integer({ min: 0, max: 100 }),
-        (initialLoadout, newLoadout, stance, yieldThreshold) => {
-          // Skip if loadout types are the same (no change)
-          if (initialLoadout === newLoadout) return true;
-
-          // Reset mock
-          mockOnRobotUpdate.mockClear();
-
-          // Create robot with initial configuration
-          const robot = createRobot(initialLoadout, stance, yieldThreshold);
-
-          render(
-            <BattleConfigTab
-              robot={robot}
-              onRobotUpdate={mockOnRobotUpdate}
-              onEquipWeapon={mockOnEquipWeapon}
-              onUnequipWeapon={mockOnUnequipWeapon}
-            />
-          );
-
-          // Find the loadout section by looking for "Weapon Loadout" text
-          const loadoutSection = screen.getByText(/Weapon Loadout/i);
-          expect(loadoutSection).toBeTruthy();
-
-          // Simulate loadout change by calling the handler directly
-          // (In a real integration test, we would click the UI element)
-          // For property-based testing, we verify the callback mechanism works
-          // by simulating what the LoadoutSelector component would do
-          mockOnRobotUpdate({ loadoutType: newLoadout });
-
-          // Property: onRobotUpdate should be called with the new loadout type
-          expect(mockOnRobotUpdate).toHaveBeenCalledWith({ loadoutType: newLoadout });
-          expect(mockOnRobotUpdate).toHaveBeenCalledTimes(1);
-
-          // Verify the callback was called with correct parameters
-          const callArgs = mockOnRobotUpdate.mock.calls[0][0];
-          expect(callArgs.loadoutType).toBe(newLoadout);
-
-          // Cleanup
-          cleanup();
-          return true;
-        }
-      ),
-      { numRuns: 100 }
-    );
-  });
-
-  it('should trigger update callback when stance changes', () => {
-    fc.assert(
-      fc.property(
-        // Generate initial stance
-        fc.constantFrom('offensive', 'defensive', 'balanced'),
-        // Generate new stance (different from initial)
-        fc.constantFrom('offensive', 'defensive', 'balanced'),
-        // Generate loadout type
-        fc.constantFrom('single', 'weapon_shield', 'two_handed', 'dual_wield'),
-        // Generate yield threshold
-        fc.integer({ min: 0, max: 100 }),
-        (initialStance, newStance, loadoutType, yieldThreshold) => {
-          // Skip if stances are the same (no change)
-          if (initialStance === newStance) return true;
-
-          // Reset mock
-          mockOnRobotUpdate.mockClear();
-
-          // Create robot with initial configuration
-          const robot = createRobot(loadoutType, initialStance, yieldThreshold);
-
-          const { rerender: _rerender } = render(
-            <BattleConfigTab
-              robot={robot}
-              onRobotUpdate={mockOnRobotUpdate}
-              onEquipWeapon={mockOnEquipWeapon}
-              onUnequipWeapon={mockOnUnequipWeapon}
-            />
-          );
-
-          // Find the stance selector section
-          const stanceSection = screen.getByText(/Battle Stance/i).parentElement;
-          expect(stanceSection).toBeTruthy();
-
-          // Simulate stance change by calling the handler
-          mockOnRobotUpdate({ stance: newStance });
-
-          // Property: onRobotUpdate should be called with the new stance
-          expect(mockOnRobotUpdate).toHaveBeenCalledWith({ stance: newStance });
-          expect(mockOnRobotUpdate).toHaveBeenCalledTimes(1);
-
-          // Verify the callback was called with correct parameters
-          const callArgs = mockOnRobotUpdate.mock.calls[0][0];
-          expect(callArgs.stance).toBe(newStance);
-
-          // Cleanup
-          cleanup();
-          return true;
-        }
-      ),
-      { numRuns: 100 }
-    );
-  });
-
-  it('should trigger update callback when yield threshold changes', () => {
-    fc.assert(
-      fc.property(
-        // Generate initial yield threshold
-        fc.integer({ min: 0, max: 100 }),
-        // Generate new yield threshold (different from initial)
-        fc.integer({ min: 0, max: 100 }),
-        // Generate loadout type
-        fc.constantFrom('single', 'weapon_shield', 'two_handed', 'dual_wield'),
-        // Generate stance
-        fc.constantFrom('offensive', 'defensive', 'balanced'),
-        (initialThreshold, newThreshold, loadoutType, stance) => {
-          // Skip if thresholds are the same (no change)
-          if (initialThreshold === newThreshold) return true;
-
-          // Reset mock
-          mockOnRobotUpdate.mockClear();
-
-          // Create robot with initial configuration
-          const robot = createRobot(loadoutType, stance, initialThreshold);
-
-          const { rerender: _rerender } = render(
-            <BattleConfigTab
-              robot={robot}
-              onRobotUpdate={mockOnRobotUpdate}
-              onEquipWeapon={mockOnEquipWeapon}
-              onUnequipWeapon={mockOnUnequipWeapon}
-            />
-          );
-
-          // Find the yield threshold section
-          const yieldSection = screen.getByText(/Yield Threshold/i).parentElement;
-          expect(yieldSection).toBeTruthy();
-
-          // Simulate yield threshold change by calling the handler
-          mockOnRobotUpdate({ yieldThreshold: newThreshold });
-
-          // Property: onRobotUpdate should be called with the new yield threshold
-          expect(mockOnRobotUpdate).toHaveBeenCalledWith({ yieldThreshold: newThreshold });
-          expect(mockOnRobotUpdate).toHaveBeenCalledTimes(1);
-
-          // Verify the callback was called with correct parameters
-          const callArgs = mockOnRobotUpdate.mock.calls[0][0];
-          expect(callArgs.yieldThreshold).toBe(newThreshold);
-
-          // Cleanup
-          cleanup();
-          return true;
-        }
-      ),
-      { numRuns: 100 }
-    );
-  });
-
-  it('should trigger update callback for any configuration change', () => {
-    fc.assert(
-      fc.property(
-        // Generate initial configuration
-        fc.record({
-          loadoutType: fc.constantFrom('single', 'weapon_shield', 'two_handed', 'dual_wield'),
-          stance: fc.constantFrom('offensive', 'defensive', 'balanced'),
-          yieldThreshold: fc.integer({ min: 0, max: 100 }),
-        }),
-        // Generate new configuration
-        fc.record({
-          loadoutType: fc.constantFrom('single', 'weapon_shield', 'two_handed', 'dual_wield'),
-          stance: fc.constantFrom('offensive', 'defensive', 'balanced'),
-          yieldThreshold: fc.integer({ min: 0, max: 100 }),
-        }),
-        (initialConfig, newConfig) => {
-          // Skip if configurations are identical (no change)
-          if (
-            initialConfig.loadoutType === newConfig.loadoutType &&
-            initialConfig.stance === newConfig.stance &&
-            initialConfig.yieldThreshold === newConfig.yieldThreshold
-          ) {
-            return true;
-          }
-
-          // Reset mock
-          mockOnRobotUpdate.mockClear();
-
-          // Create robot with initial configuration
-          const robot = createRobot(
-            initialConfig.loadoutType,
-            initialConfig.stance,
-            initialConfig.yieldThreshold
-          );
-
-          render(
-            <BattleConfigTab
-              robot={robot}
-              onRobotUpdate={mockOnRobotUpdate}
-              onEquipWeapon={mockOnEquipWeapon}
-              onUnequipWeapon={mockOnUnequipWeapon}
-            />
-          );
-
-          // Simulate configuration changes
-          if (initialConfig.loadoutType !== newConfig.loadoutType) {
-            mockOnRobotUpdate({ loadoutType: newConfig.loadoutType });
-          }
-          if (initialConfig.stance !== newConfig.stance) {
-            mockOnRobotUpdate({ stance: newConfig.stance });
-          }
-          if (initialConfig.yieldThreshold !== newConfig.yieldThreshold) {
-            mockOnRobotUpdate({ yieldThreshold: newConfig.yieldThreshold });
-          }
-
-          // Property: onRobotUpdate should be called at least once for any configuration change
-          expect(mockOnRobotUpdate).toHaveBeenCalled();
-          expect(mockOnRobotUpdate.mock.calls.length).toBeGreaterThan(0);
-
-          // Verify each call has the correct parameter structure
-          mockOnRobotUpdate.mock.calls.forEach((call) => {
-            const args = call[0];
-            expect(args).toBeDefined();
-            expect(typeof args).toBe('object');
-            
-            // Verify that at least one configuration property is present
-            const hasConfigProperty = 
-              'loadoutType' in args || 
-              'stance' in args || 
-              'yieldThreshold' in args;
-            expect(hasConfigProperty).toBe(true);
-          });
-
-          // Cleanup
-          cleanup();
-          return true;
-        }
-      ),
-      { numRuns: 100 }
-    );
-  });
-
-  it('should maintain callback consistency across multiple configuration changes', () => {
-    fc.assert(
-      fc.property(
-        // Generate a sequence of configuration changes
-        fc.array(
-          fc.record({
-            type: fc.constantFrom('loadout', 'stance', 'threshold'),
-            loadoutValue: fc.constantFrom('single', 'weapon_shield', 'two_handed', 'dual_wield'),
-            stanceValue: fc.constantFrom('offensive', 'defensive', 'balanced'),
-            thresholdValue: fc.integer({ min: 0, max: 100 }),
-          }),
-          { minLength: 1, maxLength: 5 }
-        ),
-        (changes) => {
-          // Reset mock
-          mockOnRobotUpdate.mockClear();
-
-          // Create robot with initial configuration
-          const robot = createRobot('single', 'balanced', 50);
-
-          render(
-            <BattleConfigTab
-              robot={robot}
-              onRobotUpdate={mockOnRobotUpdate}
-              onEquipWeapon={mockOnEquipWeapon}
-              onUnequipWeapon={mockOnUnequipWeapon}
-            />
-          );
-
-          // Apply each configuration change
-          changes.forEach((change) => {
-            if (change.type === 'loadout') {
-              mockOnRobotUpdate({ loadoutType: change.loadoutValue });
-            } else if (change.type === 'stance') {
-              mockOnRobotUpdate({ stance: change.stanceValue });
-            } else if (change.type === 'threshold') {
-              mockOnRobotUpdate({ yieldThreshold: change.thresholdValue });
-            }
-          });
-
-          // Property: onRobotUpdate should be called exactly as many times as there are changes
-          expect(mockOnRobotUpdate).toHaveBeenCalledTimes(changes.length);
-
-          // Verify each call corresponds to the correct change
-          changes.forEach((change, index) => {
-            const callArgs = mockOnRobotUpdate.mock.calls[index][0];
-            
-            if (change.type === 'loadout') {
-              expect(callArgs.loadoutType).toBe(change.loadoutValue);
-            } else if (change.type === 'stance') {
-              expect(callArgs.stance).toBe(change.stanceValue);
-            } else if (change.type === 'threshold') {
-              expect(callArgs.yieldThreshold).toBe(change.thresholdValue);
-            }
-          });
-
-          // Cleanup
-          cleanup();
-          return true;
-        }
-      ),
-      { numRuns: 100 }
-    );
-  });
-});
