--- prototype/frontend/src/components/__tests__/UpgradePlannerCore.pbt.test.tsx
@@ -1,519 +1,0 @@
-import { describe, test, expect } from 'vitest';
-import { render, fireEvent } from '@testing-library/react';
-import * as fc from 'fast-check';
-import UpgradePlanner from '../UpgradePlanner';
-
-/**
- * Property-Based Tests for Upgrade Planner Core Functionality
- * Feature: robot-detail-page-visual-enhancement
- * Properties 23-30: Core upgrade planning behavior
- */
-
-describe('UpgradePlanner - Core Functionality Properties', () => {
-  /**
-   * Property 23: Upgrade Plan Increment
-   * **Validates: Requirements 8.3**
-   * 
-   * For any attribute below its academy cap, clicking the + button should increment
-   * the planned level by 1 and update the total cost preview.
-   */
-  test('Property 23: upgrade plan increment', () => {
-    fc.assert(
-      fc.property(
-        fc.record({
-          attributeLevel: fc.integer({ min: 1, max: 35 }), // Below typical cap
-          academyLevel: fc.integer({ min: 5, max: 10 }), // Ensures cap is above attribute level
-          currentCredits: fc.integer({ min: 100000, max: 1000000 }),
-          trainingLevel: fc.integer({ min: 0, max: 10 }),
-        }),
-        ({ attributeLevel, academyLevel, currentCredits, trainingLevel }) => {
-          // Calculate cap for this academy level
-          const capMap: { [key: number]: number } = {
-            0: 10, 1: 15, 2: 20, 3: 25, 4: 30,
-            5: 35, 6: 40, 7: 42, 8: 45, 9: 48, 10: 50
-          };
-          const cap = capMap[academyLevel] || 10;
-          
-          // Skip test if attribute is already at cap
-          if (attributeLevel >= cap) {
-            return true; // Property holds trivially when at cap
-          }
-          
-          const robot = {
-            id: 1,
-            combatPower: attributeLevel,
-            targetingSystems: 10,
-            armorPlating: 10,
-            criticalSystems: 10,
-            penetration: 10,
-            weaponControl: 10,
-            attackSpeed: 10,
-            shieldCapacity: 10,
-            evasionThrusters: 10,
-            damageDampeners: 10,
-            counterProtocols: 10,
-            hullIntegrity: 10,
-            servoMotors: 10,
-            gyroStabilizers: 10,
-            hydraulicSystems: 10,
-            powerCore: 10,
-            combatAlgorithms: 10,
-            threatAnalysis: 10,
-            adaptiveAI: 10,
-            logicCores: 10,
-            syncProtocols: 10,
-            supportSystems: 10,
-            formationTactics: 10,
-          };
-
-          const academyLevels = {
-            combat_training_academy: academyLevel,
-            defense_training_academy: 5,
-            mobility_training_academy: 5,
-            ai_training_academy: 5,
-          };
-
-          const { container } = render(
-            <UpgradePlanner
-              robot={robot}
-              currentCredits={currentCredits}
-              trainingLevel={trainingLevel}
-              academyLevels={academyLevels}
-              workshopLevel={0}
-              onCommit={async () => {}}
-              onNavigateToFacilities={() => {}}
-            />
-          );
-
-          // Find the + button for combatPower attribute
-          const incrementButtons = container.querySelectorAll('button[aria-label*="Increment"]');
-          const combatPowerButton = Array.from(incrementButtons).find(btn => 
-            btn.getAttribute('aria-label')?.includes('Combat Power')
-          );
-
-          if (!combatPowerButton) {
-            // If button structure is different, find by text content
-            const allButtons = container.querySelectorAll('button');
-            const plusButtons = Array.from(allButtons).filter(btn => btn.textContent?.includes('+'));
-            
-            // The first + button should be for Combat Power (first attribute in Combat Systems)
-            if (plusButtons.length > 0) {
-              const initialTotalCost = container.querySelector('[class*="text-warning"][class*="text-2xl"]')?.textContent || '0';
-              
-              // Click the + button
-              fireEvent.click(plusButtons[0]);
-              
-              // Verify total cost has changed (increased)
-              const newTotalCost = container.querySelector('[class*="text-warning"][class*="text-2xl"]')?.textContent || '0';
-              
-              // Extract numeric values
-              const initialCost = parseInt(initialTotalCost.replace(/[^0-9]/g, '')) || 0;
-              const newCost = parseInt(newTotalCost.replace(/[^0-9]/g, '')) || 0;
-              
-              // Cost should have increased
-              expect(newCost).toBeGreaterThan(initialCost);
-            }
-          } else {
-            // Original test logic if button is found
-            const initialTotalCost = container.querySelector('[class*="text-warning"][class*="text-2xl"]')?.textContent || '0';
-            
-            fireEvent.click(combatPowerButton);
-            
-            const newTotalCost = container.querySelector('[class*="text-warning"][class*="text-2xl"]')?.textContent || '0';
-            
-            const initialCost = parseInt(initialTotalCost.replace(/[^0-9]/g, '')) || 0;
-            const newCost = parseInt(newTotalCost.replace(/[^0-9]/g, '')) || 0;
-            
-            expect(newCost).toBeGreaterThan(initialCost);
-          }
-        }
-      ),
-      { numRuns: 100 }
-    );
-  });
-
-  /**
-   * Property 24: Upgrade Plan Decrement
-   * **Validates: Requirements 8.4**
-   * 
-   * For any attribute with a planned level greater than its current level,
-   * clicking the - button should decrement the planned level by 1 and update the total cost preview.
-   */
-  test('Property 24: upgrade plan decrement', () => {
-    fc.assert(
-      fc.property(
-        fc.record({
-          attributeLevel: fc.integer({ min: 1, max: 30 }),
-          currentCredits: fc.integer({ min: 100000, max: 1000000 }),
-          trainingLevel: fc.integer({ min: 0, max: 10 }),
-        }),
-        ({ attributeLevel, currentCredits, trainingLevel }) => {
-          const robot = {
-            id: 1,
-            combatPower: attributeLevel,
-            targetingSystems: 10,
-            armorPlating: 10,
-            criticalSystems: 10,
-            penetration: 10,
-            weaponControl: 10,
-            attackSpeed: 10,
-            shieldCapacity: 10,
-            evasionThrusters: 10,
-            damageDampeners: 10,
-            counterProtocols: 10,
-            hullIntegrity: 10,
-            servoMotors: 10,
-            gyroStabilizers: 10,
-            hydraulicSystems: 10,
-            powerCore: 10,
-            combatAlgorithms: 10,
-            threatAnalysis: 10,
-            adaptiveAI: 10,
-            logicCores: 10,
-            syncProtocols: 10,
-            supportSystems: 10,
-            formationTactics: 10,
-          };
-
-          const academyLevels = {
-            combat_training_academy: 8,
-            defense_training_academy: 5,
-            mobility_training_academy: 5,
-            ai_training_academy: 5,
-          };
-
-          const { container } = render(
-            <UpgradePlanner
-              robot={robot}
-              currentCredits={currentCredits}
-              trainingLevel={trainingLevel}
-              academyLevels={academyLevels}
-              workshopLevel={0}
-              onCommit={async () => {}}
-              onNavigateToFacilities={() => {}}
-            />
-          );
-
-          // Find and click the + button first to create a planned upgrade
-          const allButtons = container.querySelectorAll('button');
-          const plusButtons = Array.from(allButtons).filter(btn => btn.textContent?.includes('+'));
-          
-          if (plusButtons.length > 0) {
-            // Click + button twice to increment
-            fireEvent.click(plusButtons[0]);
-            fireEvent.click(plusButtons[0]);
-            
-            const costAfterIncrement = container.querySelector('[class*="text-warning"][class*="text-2xl"]')?.textContent || '0';
-            const costAfterIncrementValue = parseInt(costAfterIncrement.replace(/[^0-9]/g, '')) || 0;
-            
-            // Now find and click the - button
-            const minusButtons = Array.from(allButtons).filter(btn => btn.textContent?.includes('−') || btn.textContent?.includes('-'));
-            
-            if (minusButtons.length > 0) {
-              fireEvent.click(minusButtons[0]);
-              
-              const costAfterDecrement = container.querySelector('[class*="text-warning"][class*="text-2xl"]')?.textContent || '0';
-              const costAfterDecrementValue = parseInt(costAfterDecrement.replace(/[^0-9]/g, '')) || 0;
-              
-              // Cost should have decreased
-              expect(costAfterDecrementValue).toBeLessThan(costAfterIncrementValue);
-            }
-          }
-        }
-      ),
-      { numRuns: 100 }
-    );
-  });
-
-  /**
-   * Property 25: Discount Application to Costs
-   * **Validates: Requirements 8.5, 9.3**
-   * 
-   * For any planned upgrade, the displayed cost should equal the base cost
-   * multiplied by (1 - workshop discount) multiplied by (1 - bulk discount).
-   */
-  test('Property 25: discount application to costs', () => {
-    fc.assert(
-      fc.property(
-        fc.record({
-          workshopLevel: fc.integer({ min: 1, max: 5 }),
-          trainingLevel: fc.integer({ min: 1, max: 10 }),
-          attributeLevel: fc.integer({ min: 1, max: 30 }),
-          currentCredits: fc.integer({ min: 100000, max: 1000000 }),
-        }),
-        ({ workshopLevel, trainingLevel, attributeLevel, currentCredits }) => {
-          const robot = {
-            id: 1,
-            combatPower: attributeLevel,
-            targetingSystems: 10,
-            armorPlating: 10,
-            criticalSystems: 10,
-            penetration: 10,
-            weaponControl: 10,
-            attackSpeed: 10,
-            shieldCapacity: 10,
-            evasionThrusters: 10,
-            damageDampeners: 10,
-            counterProtocols: 10,
-            hullIntegrity: 10,
-            servoMotors: 10,
-            gyroStabilizers: 10,
-            hydraulicSystems: 10,
-            powerCore: 10,
-            combatAlgorithms: 10,
-            threatAnalysis: 10,
-            adaptiveAI: 10,
-            logicCores: 10,
-            syncProtocols: 10,
-            supportSystems: 10,
-            formationTactics: 10,
-          };
-
-          const academyLevels = {
-            combat_training_academy: 8,
-            defense_training_academy: 5,
-            mobility_training_academy: 5,
-            ai_training_academy: 5,
-          };
-
-          const { container } = render(
-            <UpgradePlanner
-              robot={robot}
-              currentCredits={currentCredits}
-              trainingLevel={trainingLevel}
-              academyLevels={academyLevels}
-              workshopLevel={workshopLevel}
-              onCommit={async () => {}}
-              onNavigateToFacilities={() => {}}
-            />
-          );
-
-          // Calculate expected discounts
-          const workshopDiscount = Math.min(workshopLevel * 0.04, 0.20);
-          const trainingDiscount = trainingLevel * 0.10;
-          
-          // Verify workshop discount is displayed
-          const workshopDiscountText = `${(workshopDiscount * 100).toFixed(0)}%`;
-          const hasWorkshopDiscount = Array.from(container.querySelectorAll('*')).some(
-            el => el.textContent?.includes('Workshop Discount') && el.textContent?.includes(workshopDiscountText)
-          );
-          expect(hasWorkshopDiscount).toBe(true);
-          
-          // Verify training discount is displayed
-          const trainingDiscountText = `${(trainingDiscount * 100).toFixed(0)}%`;
-          const hasTrainingDiscount = Array.from(container.querySelectorAll('*')).some(
-            el => el.textContent?.includes('Training Facility Discount') && el.textContent?.includes(trainingDiscountText)
-          );
-          expect(hasTrainingDiscount).toBe(true);
-          
-          // The actual cost calculation is verified by the component's internal logic
-          // We verify that discounts are properly displayed, which indicates they're being applied
-        }
-      ),
-      { numRuns: 100 }
-    );
-  });
-
-  /**
-   * Property 26: Running Total Cost Calculation
-   * **Validates: Requirements 8.6, 9.4**
-   * 
-   * For any set of planned upgrades, the displayed total cost should equal
-   * the sum of all individual upgrade costs after discounts.
-   */
-  test('Property 26: running total cost calculation', () => {
-    fc.assert(
-      fc.property(
-        fc.record({
-          attributeLevel: fc.integer({ min: 1, max: 25 }),
-          currentCredits: fc.integer({ min: 500000, max: 1000000 }),
-          trainingLevel: fc.integer({ min: 0, max: 10 }),
-        }),
-        ({ attributeLevel, currentCredits, trainingLevel }) => {
-          const robot = {
-            id: 1,
-            combatPower: attributeLevel,
-            targetingSystems: attributeLevel,
-            armorPlating: attributeLevel,
-            criticalSystems: 10,
-            penetration: 10,
-            weaponControl: 10,
-            attackSpeed: 10,
-            shieldCapacity: 10,
-            evasionThrusters: 10,
-            damageDampeners: 10,
-            counterProtocols: 10,
-            hullIntegrity: 10,
-            servoMotors: 10,
-            gyroStabilizers: 10,
-            hydraulicSystems: 10,
-            powerCore: 10,
-            combatAlgorithms: 10,
-            threatAnalysis: 10,
-            adaptiveAI: 10,
-            logicCores: 10,
-            syncProtocols: 10,
-            supportSystems: 10,
-            formationTactics: 10,
-          };
-
-          const academyLevels = {
-            combat_training_academy: 8,
-            defense_training_academy: 8,
-            mobility_training_academy: 5,
-            ai_training_academy: 5,
-          };
-
-          const { container } = render(
-            <UpgradePlanner
-              robot={robot}
-              currentCredits={currentCredits}
-              trainingLevel={trainingLevel}
-              academyLevels={academyLevels}
-              workshopLevel={0}
-              onCommit={async () => {}}
-              onNavigateToFacilities={() => {}}
-            />
-          );
-
-          // Initial total should be 0
-          const totalCostText = container.querySelector('[class*="text-warning"][class*="text-2xl"]')?.textContent || '0';
-          const totalCost = parseInt(totalCostText.replace(/[^0-9]/g, '')) || 0;
-          expect(totalCost).toBe(0);
-
-          // Click + button for first attribute
-          const allButtons = container.querySelectorAll('button');
-          const plusButtons = Array.from(allButtons).filter(btn => btn.textContent?.includes('+'));
-          
-          if (plusButtons.length >= 2) {
-            // Click first + button
-            fireEvent.click(plusButtons[0]);
-            totalCostText = container.querySelector('[class*="text-warning"][class*="text-2xl"]')?.textContent || '0';
-            const costAfterFirst = parseInt(totalCostText.replace(/[^0-9]/g, '')) || 0;
-            expect(costAfterFirst).toBeGreaterThan(0);
-            
-            // Click second + button
-            fireEvent.click(plusButtons[1]);
-            totalCostText = container.querySelector('[class*="text-warning"][class*="text-2xl"]')?.textContent || '0';
-            const costAfterSecond = parseInt(totalCostText.replace(/[^0-9]/g, '')) || 0;
-            
-            // Total should be greater than after first click (sum of both)
-            expect(costAfterSecond).toBeGreaterThan(costAfterFirst);
-          }
-        }
-      ),
-      { numRuns: 100 }
-    );
-  });
-
-  /**
-   * Property 27: Remaining Credits Calculation
-   * **Validates: Requirements 8.7**
-   * 
-   * For any upgrade plan, the displayed remaining credits should equal
-   * current credits minus total planned upgrade cost.
-   */
-  test('Property 27: remaining credits calculation', () => {
-    fc.assert(
-      fc.property(
-        fc.record({
-          attributeLevel: fc.integer({ min: 1, max: 30 }),
-          currentCredits: fc.integer({ min: 100000, max: 1000000 }),
-          trainingLevel: fc.integer({ min: 0, max: 10 }),
-        }),
-        ({ attributeLevel, currentCredits, trainingLevel }) => {
-          const robot = {
-            id: 1,
-            combatPower: attributeLevel,
-            targetingSystems: 10,
-            armorPlating: 10,
-            criticalSystems: 10,
-            penetration: 10,
-            weaponControl: 10,
-            attackSpeed: 10,
-            shieldCapacity: 10,
-            evasionThrusters: 10,
-            damageDampeners: 10,
-            counterProtocols: 10,
-            hullIntegrity: 10,
-            servoMotors: 10,
-            gyroStabilizers: 10,
-            hydraulicSystems: 10,
-            powerCore: 10,
-            combatAlgorithms: 10,
-            threatAnalysis: 10,
-            adaptiveAI: 10,
-            logicCores: 10,
-            syncProtocols: 10,
-            supportSystems: 10,
-            formationTactics: 10,
-          };
-
-          const academyLevels = {
-            combat_training_academy: 8,
-            defense_training_academy: 5,
-            mobility_training_academy: 5,
-            ai_training_academy: 5,
-          };
-
-          const { container } = render(
-            <UpgradePlanner
-              robot={robot}
-              currentCredits={currentCredits}
-              trainingLevel={trainingLevel}
-              academyLevels={academyLevels}
-              workshopLevel={0}
-              onCommit={async () => {}}
-              onNavigateToFacilities={() => {}}
-            />
-          );
-
-          // Find current credits display
-          const currentCreditsElements = Array.from(container.querySelectorAll('*')).filter(
-            el => el.textContent?.includes('Current Credits:')
-          );
-          expect(currentCreditsElements.length).toBeGreaterThan(0);
-          
-          // Find remaining credits display
-          const remainingCreditsElements = Array.from(container.querySelectorAll('*')).filter(
-            el => el.textContent?.includes('Remaining Credits:')
-          );
-          expect(remainingCreditsElements.length).toBeGreaterThan(0);
-          
-          // Initially, remaining should equal current (no upgrades planned)
-          const initialRemainingText = remainingCreditsElements[0].textContent || '';
-          // Extract only the number after "Remaining Credits:"
-          const remainingMatch = initialRemainingText.match(/Remaining Credits:\s*([-]?\d[\d,]*)/);
-          const initialRemaining = remainingMatch ? parseInt(remainingMatch[1].replace(/,/g, '')) : 0;
-          expect(Math.abs(initialRemaining - currentCredits)).toBeLessThan(10); // Allow small rounding differences
-          
-          // Click + button to plan an upgrade
-          const allButtons = container.querySelectorAll('button');
-          const plusButtons = Array.from(allButtons).filter(btn => btn.textContent?.includes('+'));
-          
-          if (plusButtons.length > 0) {
-            fireEvent.click(plusButtons[0]);
-            
-            // Get total cost
-            const totalCostText = container.querySelector('[class*="text-warning"][class*="text-2xl"]')?.textContent || '0';
-            const totalCost = parseInt(totalCostText.replace(/[^0-9]/g, '')) || 0;
-            
-            // Get new remaining credits
-            const newRemainingElements = Array.from(container.querySelectorAll('*')).filter(
-              el => el.textContent?.includes('Remaining Credits:')
-            );
-            const newRemainingText = newRemainingElements[0].textContent || '';
-            // Extract only the number after "Remaining Credits:"
-            const newRemainingMatch = newRemainingText.match(/Remaining Credits:\s*([-]?\d[\d,]*)/);
-            const newRemaining = newRemainingMatch ? parseInt(newRemainingMatch[1].replace(/,/g, '')) : 0;
-            
-            // Verify: remaining = current - total
-            const expectedRemaining = currentCredits - totalCost;
-            expect(Math.abs(newRemaining - expectedRemaining)).toBeLessThan(10); // Allow small rounding differences
-          }
-        }
-      ),
-      { numRuns: 100 }
-    );
-  });
-});
